pyproject.toml
[project]
name = "healthcare-ml-project"
version = "0.1.0"
description = "Healthcare Data Pipeline and Machine Learning System"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "joblib>=1.3.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "apache-airflow>=2.7.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["src/healthcare_ml_project"]
